+++ src/main/java/fr/frogdevelopment/assoplus/dto/LicenceDto.java
@@ -4,28 +4,25 @@
 
 package fr.frogdevelopment.assoplus.dto;
 
-import javafx.beans.property.SimpleIntegerProperty;
-import javafx.beans.property.SimpleSetProperty;
+import javafx.beans.property.SimpleLongProperty;
 import javafx.beans.property.SimpleStringProperty;
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableSet;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 
 public class LicenceDto implements ReferenceDto {
 
-	private SimpleIntegerProperty id = new SimpleIntegerProperty();
+	private SimpleLongProperty id = new SimpleLongProperty();
 	private SimpleStringProperty code = new SimpleStringProperty();
 	private SimpleStringProperty label = new SimpleStringProperty();
 
-	public Integer getId() {
+	public Long getId() {
 		return id.get();
 	}
 
-	public SimpleIntegerProperty idProperty() {
+	public SimpleLongProperty idProperty() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id.set(id);
 	}
